Development

+++ app/Controllers/ProjectController.php
@@ -3,6 +3,7 @@
 namespace App\Controllers;
 
 use App\Controllers\BaseController;
+use App\Models\EmployeeEstimatedTimeModel;
 use App\Models\ProjectEmployeeModel;
 use App\Models\ProjectModel;
 use App\Models\RequestModel;
@@ -18,18 +19,21 @@ class ProjectController extends BaseController
     $user_obj = new UserModel();
     $project_employee_obj = new ProjectEmployeeModel();
     $request_obj = new RequestModel();
+    $employee_estimated_time_obj = new EmployeeEstimatedTimeModel();
 
     $project = $project_obj->getProjectById($id);
     $customer = $user_obj->getUserById($project['customer_id']);
     $employees_ids = $project_employee_obj->getEmployeesOfProject($id); // returns ids
     $employees = $user_obj->getUsersByIds($employees_ids);
     $requests_of_project = $request_obj->getRequestsOfProject($project['id']);
+    $time_adds = $employee_estimated_time_obj->getProjectEmployeeTimeAdds($id);
 
     return view('Project/index', [
       'project' => $project,
       'customer' => $customer,
       'employees' => $employees,
       'requests' => $requests_of_project,
+      'time_adds' => $time_adds,
       'logged_user_data' => $logged_user_data
     ]);
   }
@@ -38,11 +42,26 @@ class ProjectController extends BaseController
   {
     $logged_user_data = session()->get('logged_user');
     $project_obj = new ProjectModel();
+    $employee_estimated_time_obj = new EmployeeEstimatedTimeModel();
 
     $projects = $project_obj->getAllProjects();
+    $time_adds_before = $employee_estimated_time_obj->getAllEmployeeTimeAdds();
+    $time_adds = [];
+
+    foreach ($time_adds_before as $time_add) {
+      $el = [
+        'project_id' => $project_obj->getProjectById($time_add['project_id'])['title'],
+        'time_added' => $time_add['time_added'],
+        'created_date' => $time_add['created_date'],
+        'created_by' => $time_add['created_by'],
+      ];
+
+      array_push($time_adds, $el);
+    }
 
     return view('Project/projects', [
       'projects' => $projects,
+      'time_adds' => $time_adds,
       'logged_user_data' => $logged_user_data
     ]);
   }
@@ -54,6 +73,7 @@ class ProjectController extends BaseController
     $project_obj = new ProjectModel();
     $user_obj = new UserModel();
     $project_employee_obj = new ProjectEmployeeModel();
+    $employee_estimated_time_obj = new EmployeeEstimatedTimeModel();
 
     $project = $project_obj->getProjectById($id);
     $customer = $user_obj->getUserById($project['customer_id']);
@@ -107,8 +127,10 @@ class ProjectController extends BaseController
     $project_obj = new ProjectModel();
     if ($project_obj->edit($id, $project)) {
       $project_employee_obj = new ProjectEmployeeModel();
-      if ($project_employee_obj->setEmployeeOfProject($id, $inputedEmployees)) {
-        return redirect()->to('/project/' . $id);
+      if ($employee_estimated_time_obj->initEstimatedTime($id, $logged_user_data['id'], $estimated_time)) {
+        if ($project_employee_obj->setEmployeeOfProject($id, $inputedEmployees)) {
+          return redirect()->to('/project/' . $id);
+        }
       }
     }
 
@@ -156,12 +178,15 @@ class ProjectController extends BaseController
     }
 
     $project_obj = new ProjectModel();
+    $employee_estimated_time_obj = new EmployeeEstimatedTimeModel();
+
     if ($project_obj->create($project)) {
       $project_id = $project_obj->getProjectByTitle($project['title'])['id'];
-
-      $project_employee_obj = new ProjectEmployeeModel();
-      if ($project_employee_obj->setEmployeeOfProject($project_id, $inputedEmployees)) {
-        return redirect()->to('/projects');
+      if ($employee_estimated_time_obj->initEstimatedTime($project_id, $logged_user_data['id'], $estimated_time)) {
+        $project_employee_obj = new ProjectEmployeeModel();
+        if ($project_employee_obj->setEmployeeOfProject($project_id, $inputedEmployees)) {
+          return redirect()->to('/projects');
+        }
       }
     }
 
@@ -173,11 +198,14 @@ class ProjectController extends BaseController
     $project_obj = new ProjectModel();
     $project_employee_obj = new ProjectEmployeeModel();
     $request_obj = new RequestModel();
+    $employee_estimated_time_obj = new EmployeeEstimatedTimeModel();
 
     if ($project_obj->delete($id)) {
-      if ($project_employee_obj->deleteAllOfProject($id)) {
-        if ($request_obj->deleteAllOfProject($id)) {
-          return redirect()->to('/projects');
+      if ($employee_estimated_time_obj->deleteTimeHistoryOfProject($id)) {
+        if ($project_employee_obj->deleteAllOfProject($id)) {
+          if ($request_obj->deleteAllOfProject($id)) {
+            return redirect()->to('/projects');
+          }
         }
       }
     }
